other: split function into other files
