+++ .gitignore
@@ -29,3 +29,5 @@ public/types.js
 public/utils.js
 public/images
 .vscode/settings.json
+public/legendary_utils/library.js
+public/legendary_utils/utils.js
