+++ electron/main.ts
@@ -9,7 +9,6 @@ import {
   isLoggedIn,
   legendaryConfigPath,
   userInfo,
-  writeDefaultconfig,
   writeGameconfig,
   home,
   sidInfoUrl,
@@ -22,11 +21,8 @@ import {
   iconDark,
   getSettings,
   iconLight,
-  heroicFolder,
 } from './utils'
 
-// @ts-ignore
-import byteSize from 'byte-size'
 import { spawn, exec } from 'child_process'
 import * as path from 'path'
 import isDev from 'electron-is-dev'
@@ -34,20 +30,17 @@ import i18next from 'i18next'
 import Backend from 'i18next-fs-backend'
 
 import {
-  stat,
   readFileSync,
   writeFile,
   existsSync,
   mkdirSync,
   unlinkSync,
-  writeFileSync,
 } from 'graceful-fs'
 import { promisify } from 'util'
 import axios from 'axios'
 import { userInfo as user, cpus } from 'os'
 
 const execAsync = promisify(exec)
-const statAsync = promisify(stat)
 
 import {
   app,
@@ -58,7 +51,11 @@ import {
   Tray,
   powerSaveBlocker,
 } from 'electron'
-import { Game, InstalledInfo, KeyImage } from './types.js'
+import { Game } from './types.js'
+import {
+  getLegendaryConfig,
+  getLegendaryGames,
+} from './legendary_utils/library'
 let mainWindow: BrowserWindow = null
 
 function createWindow(): BrowserWindow {
@@ -221,38 +218,7 @@ ipcMain.on('openReleases', () => exec(`xdg-open ${heroicGithubURL}`))
 
 ipcMain.handle('checkVersion', () => checkForUpdates())
 
-ipcMain.handle('writeLibrary', async () => {
-  const { stdout, stderr } = await execAsync('legendary list-games --json')
-  if (stdout) {
-    const results = JSON.parse(stdout)
-    // NEED DOUBLE CHECK IF SOMETHINGS INSIDE IS NEEDED OR ADD MORE STUFF IT REDUCE SIZE OF THE FINAL FILE
-    // There is maybe a better way to do that tried with filter but still not work
-    results.map((res: any) => {
-      delete res.asset_info
-      delete res.metadata.creationDate
-      delete res.metadata.developerId
-      delete res.metadata.dlcItemList
-      delete res.metadata.endOfSupport
-      delete res.metadata.entitlementName
-      delete res.metadata.entitlementType
-      delete res.metadata.lastModifiedDate
-      delete res.metadata.releaseInfo
-      res.metadata.keyImages.map((k: any) => {
-        delete k.height
-        delete k.md5
-        delete k.size
-        delete k.uploadedDate
-        delete k.width
-      })
-    })
-    const json = JSON.stringify(results)
-    return writeFileSync(`${heroicFolder}library.json`, json)
-  } else if (stderr) {
-    return stderr
-  } else {
-    return 'done'
-  }
-})
+ipcMain.handle('writeLibrary', async () => getLegendaryGames())
 
 ipcMain.handle('writeFile', (event, args) => {
   const app = args[0]
@@ -488,149 +454,7 @@ ipcMain.handle('moveInstall', async (event, [appName, path]: string[]) => {
     .catch(console.log)
 })
 
-const heroicImagesFolder = `${__dirname}/images`
-
-const getCover = (appName: string, type: string, url: string) => {
-  const imagePath = `${heroicImagesFolder}/${appName}-${type}.png`
-  const imageOffline = existsSync(imagePath)
-
-  if (imageOffline) {
-    return `images/${appName}-${type}.png`
-  }
-
-  url = url.replaceAll(' ', '%20')
-
-  const downloadCommand = `curl -L '${url}' -o ${heroicImagesFolder}/${appName}-${type}.png`
-  exec(downloadCommand)
-  return `images/${appName}-${type}.png`
-}
-
-ipcMain.handle('readFile', async (event, file) => {
-  const loggedIn = isLoggedIn()
-
-  if (!isLoggedIn) {
-    return { user: { displayName: null }, library: [] }
-  }
-
-  const files: any = {
-    user: loggedIn
-      ? JSON.parse(readFileSync(userInfo, 'utf8'))
-      : { displayName: null },
-    library: `${legendaryConfigPath}/metadata/`,
-    config: heroicConfigPath,
-    installed: await statAsync(installed)
-      .then(() => JSON.parse(readFileSync(installed, 'utf-8')))
-      .catch(() => []),
-  }
-
-  if (file === 'user') {
-    if (loggedIn) {
-      writeDefaultconfig()
-      return files[file].displayName
-    }
-    return null
-  }
-
-  if (file === 'library') {
-    const library = existsSync(`${heroicFolder}library.json`)
-    const fallBackImage =
-      'https://user-images.githubusercontent.com/26871415/103480183-1fb00680-4dd3-11eb-9171-d8c4cc601fba.jpg'
-
-    if (library) {
-      return JSON.parse(readFileSync(`${heroicFolder}library.json`, 'utf-8'))
-        .map(
-          (c: {
-            metadata: {
-              description: any
-              keyImages: any
-              title: any
-              developer: any
-              customAttributes: { CloudSaveFolder: any; FolderName: any }
-            }
-            app_name: string
-            app_version: string
-            dlcs: string[]
-          }) => {
-            const {
-              description,
-              keyImages,
-              title,
-              developer,
-              customAttributes: { CloudSaveFolder, FolderName },
-            } = c.metadata
-            const app_name = c.app_name
-            const dlcs = c.dlcs
-            const cloudSaveEnabled = Boolean(CloudSaveFolder)
-            const saveFolder = cloudSaveEnabled ? CloudSaveFolder.value : ''
-            const installFolder = FolderName ? FolderName.value : ''
-            const gameBox = keyImages.filter(
-              ({ type }: KeyImage) => type === 'DieselGameBox'
-            )[0]
-            const gameBoxTall = keyImages.filter(
-              ({ type }: KeyImage) => type === 'DieselGameBoxTall'
-            )[0]
-            const logo = keyImages.filter(
-              ({ type }: KeyImage) => type === 'DieselGameBoxLogo'
-            )[0]
-
-            const art_cover = gameBox ? gameBox.url : fallBackImage
-            const art_logo = logo ? getCover(app_name, 'logo', logo.url) : null
-            const art_square = gameBoxTall ? gameBoxTall.url : fallBackImage
-
-            const cover = getCover(app_name, 'cover', art_cover)
-            const square = getCover(app_name, 'square', art_square)
-
-            const installedGames: Game[] = Object.values(files.installed)
-
-            const isInstalled = Boolean(
-              installedGames.filter((game) => game.app_name === app_name).length
-            )
-            const info = isInstalled
-              ? installedGames.filter((game) => game.app_name === app_name)[0]
-              : {}
-
-            const {
-              executable = null,
-              version = c.app_version,
-              install_size = null,
-              install_path = null,
-            } = info as InstalledInfo
-
-            const convertedSize =
-              install_size &&
-              `${byteSize(install_size).value}${byteSize(install_size).unit}`
-
-            return {
-              isInstalled,
-              info,
-              title,
-              executable,
-              version,
-              install_size: convertedSize,
-              install_path,
-              app_name,
-              developer,
-              description,
-              cloudSaveEnabled,
-              saveFolder,
-              folderName: installFolder,
-              art_cover: cover,
-              art_square: square,
-              art_logo,
-              dlcs,
-            }
-          }
-        )
-        .sort((a: { title: string }, b: { title: string }) => {
-          const gameA = a.title.toUpperCase()
-          const gameB = b.title.toUpperCase()
-          return gameA < gameB ? -1 : 1
-        })
-    }
-    return []
-  }
-  return files[file]
-})
+ipcMain.handle('readFile', async (event, file) => getLegendaryConfig(file))
 
 ipcMain.handle('egsSync', async (event, args) => {
   const linkArgs = `--enable-sync --egl-wine-prefix ${args}`
